split build graph lowering traversal
